requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pymongo==4.3.3
numpy==1.24.3
faiss-cpu==1.7.4
pillow==10.0.0
requests==2.31.0
python-multipart==0.0.6
tritonclient==2.34.0
torch==2.0.1
transformers==4.31.0
sentence-transformers==2.2.2